requirements.txt
accelerate==0.28.0
bitsandbytes==0.43.0
colorama==0.4.6
dash==2.16.1
dash-bootstrap-components==1.5.0
faster-whisper==1.0.1
gradio==4.22.0
haystack-ai==2.0.0
librosa==0.10.1
openai-whisper==20231117
pandas==2.2.1
praat-parselmouth==0.4.3
pyannote.audio==3.1.1
pyaudio==0.2.14
pydub==0.25.1
pynput==1.7.6
scipy==1.12.0
torch==2.2.1
transformers==4.38.2
vadersentiment==3.3.2
